pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fermdocs"
version = "0.1.0"
description = "Fermentation experiment document parser"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "pyyaml>=6.0",
    "pint>=0.23",
    "pandas>=2.1",
    "openpyxl>=3.1",
    "anthropic>=0.40",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.5",
    "mypy>=1.8",
    "testcontainers[postgres]>=4.0",
]
pdf = [
    "docling>=1.0",
]
gemini = [
    "google-genai>=0.3",
]

[project.scripts]
fermdocs = "fermdocs.cli:main"
fermdocs-characterize = "fermdocs_characterize.cli:main"
fermdocs-diagnose = "fermdocs_diagnose.cli:main"
fermdocs-hypothesize = "fermdocs_hypothesis.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fermdocs", "src/fermdocs_characterize", "src/fermdocs_diagnose", "src/fermdocs_hypothesis"]

[tool.hatch.build.targets.wheel.force-include]
"src/fermdocs/schema/golden_schema.yaml" = "fermdocs/schema/golden_schema.yaml"
"src/fermdocs/units/registry.txt" = "fermdocs/units/registry.txt"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
# Default test runs deselect live_llm so CI / pre-commit don't burn API
# tokens. Run explicitly with: pytest -m live_llm
addopts = "-m 'not live_llm'"
markers = [
    "live_llm: hits a real LLM API (Gemini/Anthropic). Requires API key, costs tokens. Deselected by default.",
]

[tool.mypy]
python_version = "3.11"
strict = true
